refactor(filter-chips): move trimValue + toNumber to lib/helpers
Both helpers were chip-scoped (PseudoSearchChip.logic / NumericChip.logic)
but pure generics. Move them to shared/filter-chips/lib/helpers.ts next
to the existing sanitizeFilters/chipsToFilters lib files so the pseudo-
search popover can drop its inline `String(raw ?? "").trim()` and reuse
the same helper as the chip's from-filters logic.

Addresses baz-reviewer feedback on PR #6854 (handleChange duplicating
trimValue).
